buncha utilites updates
switched debug on all commands
added face_to_vec lut in data
debug: false on items
added lib_grow_mushroom and lib_grow_fungus. They check if a mushroom or fungus can be grown before trying to grow it, which prevents the "could not generate tree at location" error the mech loves to throw
Got rid of lib_is_player and just do a simple check now since it was only every used a few times locally
added lib_is_material, which checks if a location in a direction is a material. instead of needing to do <[loc].up.material.name> == my_material all over the place, you can do <[loc].proc[is_material].context[up|dirt]> or whatever. Good for, for example, checking if a fungus is on the block it's supposed to be on, or something like that. Might remove it because it might be a bit situational but seemed useful
Added lib_is_air, which just checks if a location is an air block. I do this so often lol this one is super helpful.
Round lib_between since it throws an error if the number isn't an int
Added lib_full_volume, which gets the full volume of a cuboid, including submemebers
Added face_to_vector, which is like the opposite of vector_to_face. Takes something like "up" and turns into into a 1 length vector (0,1,0) which can be used in relative or whatever you want it for.