utilities update
Updated inject in command to use modern notation
updated use of cycle_value, since the input was swapped to be at the beginning instead of end
fixed lib_core_ease:
   uses concat now, before it used type so it always failed
   every inout case was sine, renamed them all
   updated a couple terns to use modern notation
   has permission returns false by default; if it not a player, then the thing doesn't have permission inherently
reordered clamp, cycle_value and map_range to all have input at beginning. This is because procs can be done like <[value].proc[lib_proc].context[a|b]>, so the first value has to be input to use it that way.
Added lib_lerp