all sorts of stuff
Add denchant dommand, which is an enchantment command that's a little easier to use, and more importantly, autofills in denizen register enchantments. The regular vanilla enchant command doesn't autofill, so it's hard to know if the enchantment was registered correctly and you're typing it wrong, or if the enchant never registered in the first place. Also has clever warning messages to let you know whether the enchants you are using are within vanilla range, whether the enchant will work on the item, or if the enchant will conflict with another enchant. A lot of the heavy lifiting is due to the enchant stuff monkey added for me <3
Tweaked random_placement, notable_tool usages to look a bit nicer
added dye_hex_colors, which is a map of the dye color to the dye hex code
Added vanilla enchantment_map. Might remove this, I think it's redundant as is? I'm not sure if the one thing I'm using it for I even need it
Added on item enchanted event and on prepares anvil craft to make sure enchants registered through denizen have the lore attached to them. It's not perfect if the lore already happens to contain text identical to what the enchantment is trying to do but it's serviceable, and necessary if you want denizen registered enchants to function like vanilla ones, in regards to enchanting tables and anvils. NOTE that you cant make the enchantment name pop up in the enchanting table; when you hover over a button, it'll be blank if it's a custom enchant. Nothing to do for it.
Changed some comments for clarity in core procs
Added lib ordinal; input 1 and get 1st.
Fixed weighted random to get epsilon from data script
Added one in x, a script that just returns true or false based off the number you put in, with a chance of 1/x. So if you put in 2, that's a 1/2 chance, or 50% chance you'd get a true or false.
Added mix_dyes, which takes a list of minecraft dyes and returns a colortag based on mixing those dyes together. Note that white and black dyes in vanilla are not true white and black, but in the data table this pulls from, I made it so it was true white and black, to make it easier to get the colors you want.
Added formatted date, a no context proc that just returns a nicely formatted date in the form of "3:16 PM, July 31st, 2021"