major refactor 4/15
Added a shit ton of potions that are basically vanilla potions except made as how you'd make a custom denizen item potion; that is, an awkward, non upgraded, non extended potion with a color, and then with the effect as an "additional" effect, so it's easier to get the true duration/amplifier/etc of a potion. The alternative was to put this info into a data script.