major refactor 9/15
added calculate fortune (calculates the multiplier based on vanilla's fortune values)
added apply enchant (applies an enchant to an item according to vanilla rules, unless passed defintions change behaviour. Also adds lore for denizen enchants)
added vanilla modifiers (returns the modifiers of a vanilla tool)
added custom breaking time (returns the breaking time of a block based on an item, the blocks hardness, whether a hand counts as a "valid tool" for breaking, and a map of "valid best tools". Uses vanilla logic, and returns an element of how many seconds.)
added vanilla to custom potion (pass it a vanilla potion, and get a corresponding custom denizen item that is the same as that potion effect wise, but is easier to use insofar as getting duration, color, particle info and more.)