major refactor 10/15
added cross product (vector cross product)
added dot product (vector dot product)
added plane normal (pass in three locations to define a plane, and return the normal vector of that plane)
added midpoint (returns the location that is halfway between point a and b)
added perpendicular (get a point that is perpendicular from a and b, defined by the plane of c, of an arbitrary length)
added nlerp (like lerp, but normalized?)
added slerp (not a true slerp, but rotates around a center point and goes from a to b. The closer the center point is to the midpoint of a/b, the more of a perfect circle it will be (I think? lol))
added lerp (lerps from a to b)
added rotate around (rotates around a center point, along an axis, at size distance away. percent is 0 -1, where 1 is 1 full rotation)
added get_some_rotation_points (returns a list of points from "rotate_around", for a certain density, where rate is how fast the actual loop is. This is to use a single tag to get a set of points between one location and another, rather than having to nest loops in loops in your main logic)
added ease_location (eases a location from a to b, using regular ease logic)
added to_pitch_yaw (takes a direction vector, and turns it into a vector with a mc pitch and yaw that points in the same direction the vector did)
added loc_in_box (returns true if a location is within a box defined by two locations representing the opposite corners of a box)
added rotation_snap (Returns the same location, with the pitch/yaw of the location snapped to an increment.)
added adjacent_blocks (returns a list of the adjacent blocks of a location. Does not include the corners, only blocks that share a face with this one.)
added surrounding_blocks (returns a lost of all the surrounding locations of a location)